Allow custom timeouts for AmazonMQ brokers
For peripheral AWS regions it is commong that AmazonMQ RabbitMQ for
instance, takes more than 30min for a create or even a broker update
when a cluster requires a full reboot.

Since there is already support in the code, just the API is not exposed,
this is mostly about allowing terraform code to configure the currently
static timeouts

Nothing else changes, the defaults continue to be 30m, so it is
backwards compatible.